infrastructure/requirements.txt
# =============================================================================
# Nord City Infrastructure — Unified Dependencies
# =============================================================================
# Single requirements file for ALL Python services (database, web, bot)
# and the orchestrator.  Install into one shared venv:
#
#   cd infrastructure
#   python3 -m venv venv && source venv/bin/activate
#   pip install -r requirements.txt
# =============================================================================

# ---------------------------------------------------------------------------
# Web framework & ASGI server  (database_service, web_service)
# ---------------------------------------------------------------------------
fastapi==0.104.1
uvicorn[standard]==0.24.0

# ---------------------------------------------------------------------------
# Database / ORM  (database_service; bot_service uses models only)
# ---------------------------------------------------------------------------
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# ---------------------------------------------------------------------------
# Data validation & serialization  (all services)
# ---------------------------------------------------------------------------
pydantic==2.5.2

# ---------------------------------------------------------------------------
# HTTP clients  (web_service, bot_service → database_service RPC)
# ---------------------------------------------------------------------------
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# ---------------------------------------------------------------------------
# Telegram bot  (bot_service)
# ---------------------------------------------------------------------------
python-telegram-bot==20.7

# ---------------------------------------------------------------------------
# Configuration & environment  (all services)
# ---------------------------------------------------------------------------
python-dotenv==1.0.0

# ---------------------------------------------------------------------------
# Logging  (database_service, web_service)
# ---------------------------------------------------------------------------
structlog==23.2.0

# ---------------------------------------------------------------------------
# Web extras  (web_service)
# ---------------------------------------------------------------------------
python-multipart==0.0.6
aiofiles==23.2.1

# ---------------------------------------------------------------------------
# Time utilities  (all services — shared/utils)
# ---------------------------------------------------------------------------
ntplib==0.4.0
python-dateutil==2.8.2

# ---------------------------------------------------------------------------
# Orchestrator  (rich console UI)
# ---------------------------------------------------------------------------
rich>=13.7.0
